UI changes, MapLoader asset name fix
