+++ VmodMonkeMapLoader/ComputerInterface/CommandManager.cs
@@ -1,0 +1,23 @@
+﻿using ComputerInterface;
+using Zenject;
+
+namespace VmodMonkeMapLoader.ComputerInterface
+{
+    public class CommandManager : IInitializable
+    {
+        private readonly CommandHandler _commandHandler;
+        
+        public CommandManager(CommandHandler commandHandler)
+        {
+            _commandHandler = commandHandler;
+        }
+
+        public void Initialize()
+        {
+            _commandHandler.AddCommand(new Command(name: "whoami", argumentCount: 0, args =>
+            {
+                return "MONKE";
+            }));
+        }
+    }
+}
